Extract RequestFilters component (from Requests view)
